Model: Rename `IGrid` to `Model` and move to `examples.seismic`
