+++ examples/acoustic/acoustic_example.py
@@ -2,7 +2,8 @@ import numpy as np
 
 from devito.logger import info
 from examples.acoustic.Acoustic_codegen import Acoustic_cg
-from examples.containers import IGrid, IShot
+from examples.containers import IShot
+from examples.seismic import Model
 
 
 # Velocity models
@@ -43,7 +44,7 @@ def run(dimensions=(50, 50, 50), spacing=(20.0, 20.0, 20.0), tn=1000.0,
 
     dm = 1. / (true_vp * true_vp) - 1. / (initial_vp * initial_vp)
 
-    model = IGrid(origin, spacing, true_vp, nbpml=nbpml)
+    model = Model(origin, spacing, true_vp, nbpml=nbpml)
 
     # Define seismic data.
     data = IShot()
