+++ examples/tti/tti_example.py
@@ -1,7 +1,8 @@
 import numpy as np
 
-from examples.containers import IGrid, IShot
+from examples.containers import IShot
 from examples.tti.TTI_codegen import TTI_cg
+from examples.seismic import Model
 
 
 def source(t, f0):
@@ -21,7 +22,7 @@ def setup(dimensions=(50, 50, 50), spacing=(20.0, 20.0, 20.0), tn=250.0,
     true_vp[:, :, int(dimensions[0] / 3):int(2*dimensions[0]/3)] = 3.0
     true_vp[:, :, int(2*dimensions[0] / 3):int(dimensions[0])] = 4.0
 
-    model = IGrid(origin, spacing, true_vp, nbpml=nbpml,
+    model = Model(origin, spacing, true_vp, nbpml=nbpml,
                   rho=None,
                   epsilon=.4*np.ones(dimensions),
                   delta=-.1*np.ones(dimensions),
