+++ tests/test_adjointA.py
@@ -5,7 +5,8 @@ from numpy import linalg
 from devito.logger import error
 
 from examples.acoustic.Acoustic_codegen import Acoustic_cg
-from examples.containers import IGrid, IShot
+from examples.containers import IShot
+from examples.seismic import Model
 
 
 @pytest.mark.parametrize('space_order', [4, 8, 12])
@@ -60,8 +61,8 @@ def test_acoustic(dimensions, time_order, space_order):
         error("Unknown dimension size. `dimensions` parameter"
               "must be a tuple of either size 2 or 3.")
 
-    model = IGrid(origin, spacing, true_vp, nbpml=nbpml)
-    # Define seismic data.
+    # Define seismic data
+    model = Model(origin, spacing, true_vp, nbpml=nbpml)
     data = IShot()
     src = IShot()
